all flower api added

+++ models/flower.go
@@ -4,7 +4,7 @@ type Flower struct {
 	Id int
 	Name string
   Category string
-	Price Decimal
+	Price float32
   Photo string
 	Descriptions string
 }
